calendario académico

+++ app/Http/Controllers/Academic/RoleController.php
@@ -1,39 +1,42 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Academic;
 
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\DB;
+use App\Http\Controllers\Controller;
 use App\Http\Requests\CreateRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
 use App\Models\Permission;
 use App\Models\Role;
-use App\Models\User;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\View;
 
 class RoleController extends Controller
 {
+    private $academic_period;
+
     public function __construct()
     {
         $this->middleware('auth');
         $this->middleware('check.permissions:Admin,pages.role')->only(['index', 'getAll', 'get']);
         $this->middleware('check.permissions:Admin,pages.role.modify')->only(['create', 'update']);
         $this->middleware('check.permissions:Admin,pages.role.delete')->only(['delete']);
+        $this->academic_period = View::shared('academic_period');
     }
 
     public function index()
     {
-        return view('home.role');
+        return view('access.role.index');
     }
 
     public function create(CreateRoleRequest $request)
     {
         // Verificar si el rol ya existe
-        $roleExists = Role::where('name', $request->name)->where('IsDeleted', 0)->exists();
+        $roleExists = Role::where('name', $request->name)->where('IsDeleted', 0)->where('TenantId', $this->academic_period->id)->exists();
         $permissions = $request->permissions;
         if ($roleExists) {
-            $role = Role::where('name', $request->name)->where('IsDeleted', 0)->first();
+            $role = Role::where('name', $request->name)->where('IsDeleted', 0)->where('TenantId', $this->academic_period->id)->first();
             $existing_permissions = $role->permissions()->whereIn('name', $permissions)->get();
 
             // Agregar solo los permisos que no existen
@@ -43,6 +46,7 @@ class RoleController extends Controller
             foreach ($new_permissions as $permissionName) {
                 $permissionsToSave[] = [
                     'CreatorUserId' => Auth::id(),
+                    'TenantId' => $this->academic_period->id,
                     'name' => $permissionName,
                     'roleId' => $role->id
                 ];
@@ -66,6 +70,7 @@ class RoleController extends Controller
 
         $role = new Role([
             'CreatorUserId' => Auth::id(),
+            'TenantId' => $this->academic_period->id,
             'name' => $request->name,
             'isStatic' => false,
         ]);
@@ -75,6 +80,7 @@ class RoleController extends Controller
         foreach ($permissions as $permission) {
             $permission_save = new Permission([
                 'CreatorUserId' => Auth::id(),
+                'TenantId' => $this->academic_period->id,
                 'name' => $permission,
                 'roleId' => $role->id
             ]);
@@ -91,7 +97,7 @@ class RoleController extends Controller
 
     public function delete($id)
     {
-        $role = Role::where('id', $id)->where('IsDeleted', false)->first();
+        $role = Role::where('id', $id)->where('IsDeleted', false)->where('TenantId', $this->academic_period->id)->first();
 
         if (empty($role)) {
             return response()->json([
@@ -100,7 +106,7 @@ class RoleController extends Controller
             ], 404);
         }
 
-        Permission::where('roleId', $id)->delete();
+        Permission::where('roleId', $id)->where('TenantId', $this->academic_period->id)->delete();
 
         $role->IsDeleted = true;
         $role->DeleterUserId = Auth::id();
@@ -115,7 +121,7 @@ class RoleController extends Controller
 
     public function get($id)
     {
-        $roleExist = DB::table('roles')->where('id', $id)->where('IsDeleted', false)->get();
+        $roleExist = DB::table('roles')->where('id', $id)->where('IsDeleted', false)->where('TenantId', $this->academic_period->id)->get();
 
         if (empty($roleExist)) {
             return response()->json([
@@ -124,7 +130,7 @@ class RoleController extends Controller
             ], 404);
         }
 
-        $role = Role::where('IsDeleted', false)->findOrFail($id);
+        $role = Role::where('IsDeleted', false)->where('TenantId', $this->academic_period->id)->findOrFail($id);
         $role->permissions;
 
         return response()->json([
@@ -135,7 +141,7 @@ class RoleController extends Controller
 
     public function getAll()
     {
-        $roles = Role::where('IsDeleted', false)->get();
+        $roles = Role::where('IsDeleted', false)->where('TenantId', $this->academic_period->id)->get();
 
         $data = [];
 
@@ -168,9 +174,9 @@ class RoleController extends Controller
     public function update(UpdateRoleRequest $request, $id)
     {
         // Validar si el rol ya existe
-        $roleExists = Role::where('name', $request->name)->where('IsDeleted', 0)->exists();
+        $roleExists = Role::where('name', $request->name)->where('IsDeleted', 0)->where('TenantId', $this->academic_period->id)->exists();
         $permissions = $request->permissions;
-        $roleSetting = Role::where('name', $request->name)->where('IsDeleted', 0)->first();
+        $roleSetting = Role::where('name', $request->name)->where('IsDeleted', 0)->where('TenantId', $this->academic_period->id)->first();
 
         if ($roleExists && $roleSetting->id != $id) {
             $existing_permissions = $roleSetting->permissions()->whereIn('name', $permissions)->get();
@@ -183,6 +189,7 @@ class RoleController extends Controller
             foreach ($new_permissions as $permissionName) {
                 $permissionsToSave[] = [
                     'CreatorUserId' => Auth::id(),
+                    'TenantId' => $this->academic_period->id,
                     'name' => $permissionName,
                     'roleId' => $roleSetting->id
                 ];
@@ -207,7 +214,7 @@ class RoleController extends Controller
             ], 200);
         }
 
-        $role = Role::where('id', $id)->where('IsDeleted', false)->first();
+        $role = Role::where('id', $id)->where('IsDeleted', false)->where('TenantId', $this->academic_period->id)->first();
 
         // Verificar si el rol no existe
         if (empty($role)) {
@@ -232,6 +239,8 @@ class RoleController extends Controller
         // Agregar los nuevos permisos
         foreach ($new_permissions as $permissionName) {
             Permission::create([
+                'CreatorUserId' => Auth::id(),
+                'TenantId' => $this->academic_period->id,
                 'name' => $permissionName,
                 'roleId' => $role->id
             ]);
@@ -240,7 +249,7 @@ class RoleController extends Controller
         // Actualizar los nombres de los permisos modificados
         $permissionsToUpdate = array_intersect($currentPermissions, $permissions);
         foreach ($permissionsToUpdate as $permissionName) {
-            $existingPermission = $role->permissions->where('name', $permissionName)->first();
+            $existingPermission = $role->permissions->where('name', $permissionName)->where('TenantId', $this->academic_period->id)->first();
             $existingPermission->update(['name' => $permissionName]);
         }
 
@@ -263,7 +272,7 @@ class RoleController extends Controller
     public function revoke(Request $request, $id)
     {
         // Verificar si el rol ya existe
-        $role = Role::where('id', $id)->where('IsDeleted', false)->first();
+        $role = Role::where('id', $id)->where('IsDeleted', false)->where('TenantId', $this->academic_period->id)->first();
 
         if (empty($role)) {
             return response()->json([
@@ -278,7 +287,7 @@ class RoleController extends Controller
 
         foreach ($role_permissions as $role_permission) {
             if ($role_permission->name == $request->name) {
-                Permission::where('name', $request->name)->where('roleId', $id)->delete();
+                Permission::where('name', $request->name)->where('roleId', $id)->where('TenantId', $this->academic_period->id)->delete();
                 $permissionExists = true;
                 break;
             }
